pyproject.toml
[project]
name = "agentic-graph-rag"
version = "0.1.0"
description = "Agentic Graph RAG with Neo4j"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "neo4j>=5.0",
    "qdrant-client==1.13",
    "openai>=1.0",
    "prompt-toolkit>=3.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "datasets>=4.5.0",
]

[project.scripts]
agentic-graph-rag = "agentic_graph_rag.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-anyio>=0.0.0",
    "pytest-cov>=4.1.0",
    "python-dotenv>=1.0",
    "ruff>=0.0.0",
    "pyrefly>=0.0.0",
    "pre-commit>=3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_graph_rag"]

[tool.pytest.ini_options]
# Integration tests require live services — skip them by default.
# Run explicitly with: uv run pytest -m integration
addopts = "-m 'not integration'"
markers = [
    "integration: requires live external services (Neo4j, OpenAI, etc.)",
]

[tool.pyrefly]
project-includes = [
    "**/*.py*",
    "**/*.ipynb",
]

[tool.coverage.run]
source = ["src/agentic_graph_rag"]
omit = ["src/agentic_graph_rag/scripts/*"]
